Extract binary upload load test into parameterized
